+++ packages/docusaurus-plugin-content-pages/index.js
@@ -7,14 +7,12 @@
 
 const globby = require('globby');
 const path = require('path');
-
-// TODO: Do not make it relative because plugins can be from node_modules.
 const {encodePath, fileToPath, idx} = require('@docusaurus/utils');
 
 const DEFAULT_OPTIONS = {
   metadataKey: 'pagesMetadata',
   metadataFileName: 'pagesMetadata.json',
-  path: 'pages', // Path to data on filesystem.
+  path: 'pages', // Path to data on filesystem, relative to site dir.
   routeBasePath: '', // URL Route.
   include: ['**/*.{js,jsx}'], // Extensions to include.
   component: '@theme/Pages',
@@ -29,6 +27,10 @@ class DocusaurusPluginContentPages {
 
   getName() {
     return 'docusaurus-plugin-content-pages';
+  }
+
+  getPathsToWatch() {
+    return [this.contentPath];
   }
 
   async loadContent() {
@@ -101,10 +103,6 @@ class DocusaurusPluginContentPages {
         modules: [source],
       });
     });
-  }
-
-  getPathsToWatch() {
-    return [this.contentPath];
   }
 }