+++ packages/docusaurus/CHANGES.md
@@ -1,3 +1,8 @@
 # Breaking Changes
 
+### `siteConfig.js` changes
+
 - `siteConfig.js` renamed to `docusaurus.config.js`.
+- Removed the following config options:
+  - `docsUrl`. Use the plugin option on `docusaurus-plugin-content-blog` instead.
+  - `customDocsPath`. Use the plugin option on `docusaurus-plugin-content-blog` instead.